feat: refactor lightbox
